requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
Django==5.2.5
django-allauth==65.11.0
idna==3.10
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-decouple==3.8
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
